refactor(handlers): extract guest handlers (WP5.9.5.2)
Extract guest-related handlers into separate useGuestHandlers hook:
- validateGuestName
- handleToggleGuestForm
- handleGuestNameChange
- handleAddGuest

useRegistrationHandlers.js: 1,156 → 993 lines (-163 lines)
guestHandlers.js: 239 lines (new)
