+++ src/registration/appHandlers/handlers/index.js
@@ -1,1 +1,2 @@
 export { useAdminHandlers } from './adminHandlers.js';
+export { useGuestHandlers } from './guestHandlers.js';
